Add a ton of documentation and pull things into and out of files as appropriate
