--- Sources/ApolloCodegenLib/ASTArgument.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-/// An argument which can be 
-class ASTArgument: Codable {
-  let name: String
-  let value: JSONContainer
-  let type: ASTVariableType
-}
